+++ CHANGELOG.md
@@ -17,6 +17,10 @@ Version 0.7.1.0 (compiler 0.7.1.0, engine 0.7.3.0):
 - Compiler:
   - Fixed a bug that disallowed the comment character (#) inside strings.
   - Added allowing of a comment following the line continuation character (\\).
+  - Enhanced the 'for' statement to allow tuple assignment of arbitrary depth,
+    like with assignments.
+  - Fixed a code generation bug with tuple constants whose first element is
+    a tuple constant (in parentheses).
 
 Version 0.7.0.0 (compiler 0.7.0.0, engine 0.7.0.0):
 - Initial release.